Separate torch_cuda_utils.h
Some of the functions do not need CUDA device.
Move them to torch_utils.h

Update loading-pt cpp file so loading-pt example does not require GPU now.
